requirements.txt
PyQt6==6.7.0
sqlalchemy==2.0.23
alembic==1.13.1
openai>=1.50.0
mistralai==0.4.2
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
pydantic==2.5.2
pydantic-settings==2.1.0
openpyxl==3.1.2
python-dotenv==1.0.0
asyncio-throttle==1.0.2

